Return stack trace on import errors
This commit calls the methodname method which is added to the
ModuleImportFailure testSuite class when an import failure occurs.
That method just raises an ImportError with the original failures
traceback as the message. So by appending the exception string to
the error list it makes debugging the import failure much easier.